wave: : use BOARD_EGL_SYSTEMUI_PBSIZE_HACK
Here we adjust the PB size from its default value to 512KB which
is the minimum acceptable for the systemui process.
We do this on low-end devices only because it allows us to enable
h/w acceleration in the systemui process while keeping the
memory usage down.

Obviously, this is the wrong way to make this
adjustment, but at the time of this writing this was the safest
solution.

requires: https://gerrit.omnirom.org/12531
Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: I1608080a5fd11bd4d76352a1fe0ff8486c374bf6